[Refactor] STAMPIT-117 - AuthRepo-AccountAmangeRepo 분리
1. 마이페이지에서 사용하는 그룹탈퇴, 계정탈퇴, 로그아웃 분리 작업 및 Manager DI 교체
