+++ StampIt-Project/StampIt-Project/Domain/Repository/AccountManageRepository.swift
@@ -1,0 +1,16 @@
+//
+//  AccountManageRepository.swift
+//  StampIt-Project
+//
+//  Created by iOS study on 6/24/25.
+//
+
+import Foundation
+import RxSwift
+
+protocol AccountManageRepositoryProtocol {
+    func signOut() -> Observable<Void>
+    func deleteAccount() -> Observable<Void>
+    func getGroupMemberCount(groupId: String) -> Observable<Int>
+    func leaveGroup() -> Observable<User>
+}
