+++ StampIt-Project/StampIt-Project/Domain/Repository/AuthRepository.swift
@@ -35,13 +35,8 @@ protocol AuthRepositoryProtocol {
         invite: Invitation
     ) -> Observable<StampIt_Project.User>
     
-    // MARK: - 계정 관리
-    func signOut() -> Observable<Void>
-    func deleteAccount() -> Observable<Void>
-    func leaveGroup() -> Observable<User>
-    func getGroupMemberCount(groupId: String) -> Observable<Int>
-    
     // MARK: - 편의 메서드 (Extension에서 구현된 것들)
     func getCurrentGroupID() -> Observable<String>
     func getCurrentUserID() -> Observable<String>
-    func isCurrentUserLeader() -> Observable<Bool>}
+    func isCurrentUserLeader() -> Observable<Bool>
+}
